backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for calculus done right"
authors = ["Ulisse Mini <ulisse.mini@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.1"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
python-dotenv = "^0.19.2"
tortoise-orm = {extras = ["asyncpg"], version = "^0.18.0"}
aiohttp = "^3.8.1"
aiohttp-requests = "^0.1.3"
google-auth = "^2.3.3"
six = "^1.16.0"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
